--- Sleep Tracker/SleepTracker/Helpers/AppFontSize.swift
@@ -1,11 +1,0 @@
-import SwiftUI
-
-/// The font size used in the application.
-enum AppFontSize {
-    /// A font size used for small details.
-    static let detail: CGFloat = 12
-    /// A base font size.
-    static let base: CGFloat = 20
-    /// A font size used for headings.
-    static let heading: CGFloat = 40
-}
