+++ Sleep Tracker/SleepTracker/Helpers/CGFloat+AppFontSize.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+extension CGFloat {
+    /// The font size used in the application.
+    enum AppFontSize {
+        /// A font size used for small text elements.
+        static let small: CGFloat = 12
+        /// A base font size used as the default font size.
+        static let base: CGFloat = 20
+        /// A font size used for headings.
+        static let heading: CGFloat = 40
+    }
+}
