+++ Sleep Tracker/SleepTrackerTests/Helpers/AppUIConstantTests.swift
@@ -8,11 +8,11 @@ final class AppUIConstantTests: XCTestCase {
         let largeCornerRadius: CGFloat = 80
 
         let transparent = 0.0
-        let translucent = 0.5
+        let semitransparent = 0.5
         let opaque = 1.0
 
-        let minFrameWidth: CGFloat = 0
-        let maxFrameHeight: CGFloat = 100
+        let minFrameWidthInAlarmSetup: CGFloat = 0
+        let maxFrameHeightInAlarmSetup: CGFloat = 100
 
         XCTAssertEqual(baseCornerRadius, AppUIConstant.baseCornerRadius)
         XCTAssertEqual(alarmCornerRadius, AppUIConstant.alarmCornerRadius)
@@ -20,10 +20,10 @@ final class AppUIConstantTests: XCTestCase {
         XCTAssertEqual(largeCornerRadius, AppUIConstant.largeCornerRadius)
 
         XCTAssertEqual(transparent, AppUIConstant.transparent)
-        XCTAssertEqual(translucent, AppUIConstant.translucent)
+        XCTAssertEqual(semitransparent, AppUIConstant.semitransparent)
         XCTAssertEqual(opaque, AppUIConstant.opaque)
 
-        XCTAssertEqual(minFrameWidth, AppUIConstant.minFrameWidth)
-        XCTAssertEqual(maxFrameHeight, AppUIConstant.maxFrameHeight)
+        XCTAssertEqual(minFrameWidthInAlarmSetup, AppUIConstant.minFrameWidthInAlarmSetup)
+        XCTAssertEqual(maxFrameHeightInAlarmSetup, AppUIConstant.maxFrameHeightInAlarmSetup)
     }
 }
